refactor base clases. added exercises endpoints

+++ src/main/java/com/home/main_service/person/UpdatePersonDTO.java
@@ -1,6 +1,6 @@
 package com.home.main_service.person;
 
-import com.home.main_service.base.BaseUpdateDTO;
+import com.home.main_service.base.globalbase.BaseUpdateDTO;
 import lombok.Data;
 
 import java.time.Instant;
@@ -8,7 +8,7 @@ import java.time.Instant;
 @Data
 public class UpdatePersonDTO extends BaseUpdateDTO {
     private Gender gender;
-    private double height;
-    private double weight;
+    private int height;
+    private float weight;
     private Instant birthdate;
 }
